Feature: Add support for provide base image
Adds two new options `base-image` and `public-image`.

`base-image` supplies the name of the base image to use when creating
an image to push. The image is pulled for you. This avoids having to
know the layer id of the primary image being pulled and updated as
part of the tests.

`public-image` is a simple flag to indicate if the base image specified
is public (true) or private (false) (default is false). If private the
credentials provided will be used when pulling the base image.

The change is completely backwards-compatiable with previous options.