Feature: Minimal image for runtime
Leverage a dev image for building the monitor binary and then copy
the binary into an alpine image for use at runtime.

Uses glide to manage dependencies so that each build will be repeatable.
Also includes a Makefile to assist with adding dependencies, creating
the local vendor directory, building the binary, and building the image.